tools: ceph-monstore-tool: generate mdsmaps on-demand
This patch allows one to generate mdsmaps on a mon store while keeping
the store sane and consistent.  Usage of this tool requires that no
monitor is currently using the specified mon store data path, or any
other tool that may use leveldb.

DISCLAIMER: This patch is to be used for testing purposes only.  This is
not a piece of code that should make into the main tree as it would need
tidying up and, considering it would be useful for other monitor maps as
well, in that case the least we could do would be to add such support.
The existence of this patch arouse from the need to test mdsmap trimming
on a large enough store.  Also, there's a chance that leveldb may crash
due to depletion of fds after the maps are generated, due to a failed
attempt at compaction, if the number of generated maps is big enough.


Usage:

  ceph-monstore-tool \
      --mon-store-path </path/to/ceph-mon.X/data> \
      --command gen-mdsmaps

or

  ceph-monstore-tool \
      --mon-store-path </path/to/ceph-mon.X/data> \
      --command gen-mdsmaps \
      --version N

where N can be any integer value greater than 1.  If '--version N' is
omitted, we will generate 1000 maps by default.


Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>